backend_django/requirements.txt
#requirements for python environment
Django==4.1.4
djangorestframework>=3.12,<4.0
dj-database-url>=0.5.0
psycopg2-binary>=2.8.6
django-cors-headers>=3.7.0
pillow
django-cors-headers
PyJWT
django-filter
python-dotenv
djangorestframework_simplejwt
requests
django-environ